Тотальная вырезня
Удаление рендеров из аномали (https://www.youtube.com/watch?v=rLENDOmAipI)
Удалена команда language (ломала всё подряд)

+++ src/Layers/xrRenderPC_R1/LightProjector.cpp
@@ -254,9 +254,9 @@ void CLightProjector::calculate	()
 		RCache.set_xform_view	(mView);
 
 		// Select slot, set viewport
-		u32 s_x = c_it%P_o_line;
-		u32 s_y = c_it/P_o_line;
-		D3DVIEWPORT9 VP = {s_x*P_o_size,s_y*P_o_size,P_o_size,P_o_size,0,1 };
+		DWORD s_x =	static_cast<DWORD>(c_it%P_o_line);
+		DWORD s_y =	static_cast<DWORD>(c_it/P_o_line);
+		D3DVIEWPORT9 VP			=	{s_x*P_o_size,s_y*P_o_size,P_o_size,P_o_size,0,1 };
 		CHK_DX					(HW.pDevice->SetViewport(&VP));
 
 		// Clear color to ambience
